+++ tests/broken_links/src/lib.rs
@@ -1,1 +1,2 @@
 //! Links to [not here](./fn.not_here.html)
+//! with [intra-doc](links)